+++ kbound/__init__.py
@@ -18,7 +18,7 @@ Quickstart:
     16
 """
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"
 
 from kbound.recover import recover_rule, RecoveryResult
 from kbound.classifier.oracle_classifier import classify_geometry